orchestration::actions::action

Trait ActionTrait

pub trait ActionTrait: Send {
    // Required methods
    fn execute(&mut self) -> ActionFuture;
    fn name(&self) -> &'static str;
    fn dbg_fmt(&self, nest: usize, f: &mut Formatter<'_>) -> Result;
    fn fill_runtime_info(&mut self, p: &mut ActionRuntimeInfoProvider);
}

Describes action interface that let us build task chain from program. Each action should store it’s actions as Box<dyn ActionTrait> for now

Required Methods

fn execute(&mut self) -> ActionFuture

Will be called on each Program iteration.

Key assumptions: - should avoid allocation except creation of boxed future - each action shall propagate ActionResult down the chain in Future and should immediately stop it’s work once Err(_) is reached, propagating it down.

fn name(&self) -> &'static str

Provide debug name of action

fn dbg_fmt(&self, nest: usize, f: &mut Formatter<'_>) -> Result

Since we store actions behind dyn ActionTrait, we need an API that we can call from program to print constructed representation

fn fill_runtime_info(&mut self, p: &mut ActionRuntimeInfoProvider)